+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `ActionBar` - Added a display for the current application version.
+- `ErrorHandling` - Added a global error boundary component to catch and display errors gracefully to the user.
 
 ### Changed