+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/Synogun/GraphVM/compare/v2.4.0...develop)
+## [Unreleased](https://github.com/Synogun/GraphVM/compare/v2.5.0...develop)
+
+### Added
+
+### Changed
+
+### Fixed
+
+## [v2.5.0](https://github.com/Synogun/GraphVM/compare/v2.4.0...v2.5.0)
 
 ### Added